Buildfixes, dependency shuffling to make things more logical
