+++ libretro/CMakeLists.txt
@@ -42,4 +42,7 @@ if(NOT MSVC)
    	target_link_libraries(ppsspp_libretro "-Wl,-Bsymbolic")
    endif()
 endif()
-target_link_libraries(ppsspp_libretro ${LinkCommon})
+target_link_libraries(ppsspp_libretro core ${CMAKE_THREAD_LIBS_INIT})
+if(ANDROID)
+	target_link_libraries(ppsspp_libretro log)
+endif()
